synchronized_value vocabulary type
The synchronized_value type is useful for binding a value with a mutex
and reducing the chance of forgetting to lock a mutex. While it does
not fully prevent deadlocks, it does have an apply() method which
allows gaining synchronized access to multiple synchronized_values at
the same time, by using a deadlock avoidance algorithm based on
try_lock().

Signed-off-by: Damian Jarek <damian.jarek93@gmail.com>